+++ .erb/configs/webpack.config.renderer.prod.babel.js
@@ -30,7 +30,7 @@ export default merge(baseConfig, {
   entry: [
     'core-js',
     'regenerator-runtime/runtime',
-    path.join(__dirname, '../../src/index.tsx'),
+    path.join(__dirname, '../../src/client/index.tsx'),
   ],
 
   output: {
