+++ src/client/page/index/index.css
@@ -92,3 +92,7 @@
   width: 200px;
   height: 200px;
 }
+
+.font10 {
+  font-size: 10px;
+}
