+++ src/shared/utils.js
@@ -4,4 +4,8 @@ export function takeDirectory() {
   return ipcRenderer.invoke('selectDirectory');
 }
 
+export function getConfig() {
+  return ipcRenderer.invoke('getConfig');
+}
+
 export const unused = 1;
